p1: DEBUGGING Harkleswarm.c
I've been troubleshooting a (ONE?) BUG in shwarm_it.c so I've been
adding tons of DEBUGGING statements.

GOOD NEWS: I think I know the problem. On a line formed by points A, B,
and C (in that order), point C is attempting to move in between points A
and B.  On its way, it overlaps point B and Harkleswarm core dumps the
bed.

BAD NEWS: After staring at it for a while, I think the solution is to
refactor find_closest_points() into an entry level wrapper function that
calls tailored find_closest_points_one_dim() (and such).  I don't think
that this find_closest_points() function is the "one size fits all" I
planned it to be.  I need the ability to find the nearest points... in
dimensionally appropriate ways.  For instance, in one dimension... I
need to find the nearest point "up slope" and "down slope" (if those are
even official terms).  For two dimensions, I need to find the three
nearest points that form a triangle encapsulating the current point.

Changes to be committed:
	modified:   Harkleswarm.c