p2: Big Harkleswarm.c Change (local prototypes)
The local functions were getting unwieldy so I moved all the local
function definitions to the end of the source file and added function
prototypes to the top.

The library compiles but this is an interim commit.  I'm still working
on fixing some problems with my triangle calculations.

Changes to be committed:
	modified:   Harkleswarm.c