p2: EUREKA!
Everything appears to be working.
The equilibrium isn't quite what I envisioned because it's using
triangles but it *IS* a triangular equilibrium.
Maybe I should transition it to an equidistant equilibrium:
	1. Find the four closest encapsulating points
	2. Find the center (centroid?) of that geometric shape
	3. Move there
Maybe that's another phase.  For now, I'm going to ramp up the number of
points and watch the behavior.

There's a lot of stderr DEBUGGING still taking place so be sure to
redeirect stderr to a file or /dev/null.

Changes to be committed:
	modified:   shwarm_it.c